Moving spatial information to visual peception
The system currently stores an agent's location information as the state of that object (which actually complicated the declaration of some affordances). But after some thought I felt it would be more intuitive if the spatial information is part of visual perception. So, I moved all the spatial information to visual memory of the agent.

- Implementation to save agent memories as memory events instead of string traces

- Modified .gitignore

- Changes to store visual information in a MemoryEvent

- Added two memory types for event starting and ending